refactor: update images as per theming
