+++ src/Shared/Components/CICDHistory/CDEmptyState.tsx
@@ -16,7 +16,7 @@
 
 import React from 'react'
 import { GenericEmptyState, ImageType } from '../../../Common'
-import AppNotDeployed from '../../../Assets/Img/app-not-deployed.png'
+import AppNotDeployed from '../../../Assets/Img/app-not-deployed.svg'
 import { EMPTY_STATE_STATUS } from '../../constants'
 
 const CDEmptyState = ({
